admin: reorganised files into packages
