--- src/main/java/Exceptions/ParseErrors/UnknownCommandException.java
@@ -1,10 +1,0 @@
-package Exceptions.ParseErrors;
-
-import Exceptions.ParseException;
-
-public class UnknownCommandException extends ParseException {
-    public UnknownCommandException(String cmd){
-        super(cmd);
-    }
-
-}
