--- src/main/java/Exceptions/ParseException.java
@@ -1,7 +1,0 @@
-package Exceptions;
-
-public class ParseException extends BotException {
-    public ParseException(String msg){
-        super(ErrorCode.PARSE_ERROR, msg);
-    }
-}
