--- src/main/java/Exceptions/StorageErrors/StorageFileNotFoundException.java
@@ -1,9 +1,0 @@
-package Exceptions.StorageErrors;
-
-import Exceptions.StorageException;
-
-public class StorageFileNotFoundException extends StorageException {
-    public StorageFileNotFoundException(String msg) {
-        super(msg);
-    }
-}
