--- src/main/java/Exceptions/StorageErrors/StorageFormatException.java
@@ -1,9 +1,0 @@
-package Exceptions.StorageErrors;
-
-import Exceptions.StorageException;
-
-public class StorageFormatException extends StorageException {
-    public StorageFormatException(String msg){
-        super(msg);
-    }
-}
