--- src/main/java/Exceptions/StorageException.java
@@ -1,7 +1,0 @@
-package Exceptions;
-
-public class StorageException extends BotException{
-    public StorageException(String msg) {
-        super(ErrorCode.STORAGE_ERROR, msg);
-    }
-}
