--- src/main/java/Exceptions/ValidationErrors/FileNotFoundException.java
@@ -1,9 +1,0 @@
-package Exceptions.ValidationErrors;
-
-import Exceptions.ValidationException;
-
-public class FileNotFoundException extends ValidationException {
-    public FileNotFoundException(String details) {
-        super(details);
-    }
-}
