--- src/main/java/Exceptions/ValidationErrors/IllegalArgumentException.java
@@ -1,9 +1,0 @@
-package Exceptions.ValidationErrors;
-
-import Exceptions.ValidationException;
-
-public class IllegalArgumentException extends ValidationException {
-    public IllegalArgumentException(String what) {
-        super("Illegal Argument- " + what + ".");
-    }
-}
