--- src/main/java/Exceptions/ValidationErrors/IndexOutOfBoundsException.java
@@ -1,9 +1,0 @@
-package Exceptions.ValidationErrors;
-
-import Exceptions.ValidationException;
-
-public class IndexOutOfBoundsException extends ValidationException {
-    public IndexOutOfBoundsException(int index) {
-        super("Index [" + index + "] is out of range.");
-    }
-}
