--- src/main/java/Exceptions/ValidationErrors/MissingArgumentException.java
@@ -1,9 +1,0 @@
-package Exceptions.ValidationErrors;
-
-import Exceptions.ValidationException;
-
-public class MissingArgumentException extends ValidationException {
-    public MissingArgumentException(String what) {
-        super("Missing " + what + "!!");
-    }
-}
