--- src/main/java/Exceptions/ValidationException.java
@@ -1,7 +1,0 @@
-package Exceptions;
-
-public class ValidationException extends BotException {
-    public ValidationException(String msg) {
-        super(ErrorCode.VALIDATION_ERROR, msg);
-    }
-}
