--- src/main/java/StaticConfig.java
@@ -1,11 +1,0 @@
-public final class StaticConfig {
-    public static final int MAX_TASKS = 100;    // sets the hard limit for TaskList
-    public static final String APP_NAME = "Golden";
-    public static final String INIT_FILEPATH = "src/main/data/test1.txt";
-}
-
-/*
-* This class defines all the constants of the Chatbot app.
-* static = the attributes are shared across all instances
-* final = the values cannot be changed
-* */
