+++ src/main/java/com/golden/exceptions/ParseException.java
@@ -1,0 +1,7 @@
+package com.golden.exceptions;
+
+public class ParseException extends BotException {
+    public ParseException(String msg){
+        super(ErrorCode.PARSE_ERROR, msg);
+    }
+}
