+++ src/main/java/com/golden/exceptions/StorageException.java
@@ -1,0 +1,7 @@
+package com.golden.exceptions;
+
+public class StorageException extends BotException{
+    public StorageException(String msg) {
+        super(ErrorCode.STORAGE_ERROR, msg);
+    }
+}
