+++ src/main/java/com/golden/exceptions/ValidationException.java
@@ -1,0 +1,7 @@
+package com.golden.exceptions;
+
+public class ValidationException extends BotException {
+    public ValidationException(String msg) {
+        super(ErrorCode.VALIDATION_ERROR, msg);
+    }
+}
