+++ src/main/java/com/golden/exceptions/parseErrors/UnknownCommandException.java
@@ -1,0 +1,11 @@
+package com.golden.exceptions.parseErrors;
+
+
+import com.golden.exceptions.ParseException;
+
+public class UnknownCommandException extends ParseException {
+    public UnknownCommandException(String cmd){
+        super(cmd);
+    }
+
+}
