+++ src/main/java/com/golden/exceptions/storageErrors/StorageFileNotFoundException.java
@@ -1,0 +1,9 @@
+package com.golden.exceptions.storageErrors;
+
+import com.golden.exceptions.StorageException;
+
+public class StorageFileNotFoundException extends StorageException {
+    public StorageFileNotFoundException(String msg) {
+        super(msg);
+    }
+}
