+++ src/main/java/com/golden/exceptions/storageErrors/StorageFormatException.java
@@ -1,0 +1,9 @@
+package com.golden.exceptions.storageErrors;
+
+import com.golden.exceptions.StorageException;
+
+public class StorageFormatException extends StorageException {
+    public StorageFormatException(String msg){
+        super(msg);
+    }
+}
