+++ src/main/java/com/golden/exceptions/validationErrors/FileNotFoundException.java
@@ -1,0 +1,9 @@
+package com.golden.exceptions.validationErrors;
+
+import com.golden.exceptions.ValidationException;
+
+public class FileNotFoundException extends ValidationException {
+    public FileNotFoundException(String details) {
+        super(details);
+    }
+}
