+++ src/main/java/com/golden/exceptions/validationErrors/IllegalArgumentException.java
@@ -1,0 +1,9 @@
+package com.golden.exceptions.validationErrors;
+
+import com.golden.exceptions.ValidationException;
+
+public class IllegalArgumentException extends ValidationException {
+    public IllegalArgumentException(String what) {
+        super("Illegal Argument- " + what + ".");
+    }
+}
