+++ src/main/java/com/golden/exceptions/validationErrors/IndexOutOfBoundsException.java
@@ -1,0 +1,9 @@
+package com.golden.exceptions.validationErrors;
+
+import com.golden.exceptions.ValidationException;
+
+public class IndexOutOfBoundsException extends ValidationException {
+    public IndexOutOfBoundsException(int index) {
+        super("Index [" + index + "] is out of range.");
+    }
+}
