+++ src/main/java/com/golden/exceptions/validationErrors/MissingArgumentException.java
@@ -1,0 +1,9 @@
+package com.golden.exceptions.validationErrors;
+
+import com.golden.exceptions.ValidationException;
+
+public class MissingArgumentException extends ValidationException {
+    public MissingArgumentException(String what) {
+        super("Missing " + what + "!!");
+    }
+}
